Add side by side to Cesium 3D Tiles example
